+++ lib/action_cable/connection/base.rb
@@ -133,7 +133,7 @@ module ActionCable
           type: ActionCable::INTERNAL[:message_types][:disconnect],
           reason: reason,
           reconnect: reconnect
-        )
+        ) rescue nil
         socket.close
       end